+++ pkg/util/push/push.go
@@ -9,7 +9,6 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/exp/api/remote"
 	"github.com/prometheus/prometheus/model/labels"
-	writev2 "github.com/prometheus/prometheus/prompb/io/prometheus/write/v2"
 	"github.com/prometheus/prometheus/util/compression"
 	"github.com/weaveworks/common/httpgrpc"
 	"github.com/weaveworks/common/middleware"
@@ -79,12 +78,17 @@ func Handler(remoteWrite2Enabled bool, maxRecvMsgSize int, sourceIPs *middleware
 		}
 
 		handlePRW2 := func() {
-			var req writev2.Request
+			var req cortexpb.PreallocWriteRequestV2
 			err := util.ParseProtoReader(ctx, r.Body, int(r.ContentLength), maxRecvMsgSize, &req, util.RawSnappy)
 			if err != nil {
 				level.Error(logger).Log("err", err.Error())
 				http.Error(w, err.Error(), http.StatusBadRequest)
 				return
+			}
+
+			req.SkipLabelNameValidation = false
+			if req.Source == 0 {
+				req.Source = cortexpb.API
 			}
 
 			v1Req, err := convertV2RequestToV1(&req)
@@ -165,7 +169,7 @@ func setPRW2RespHeader(w http.ResponseWriter, samples, histograms, exemplars int
 	w.Header().Set(rw20WrittenExemplarsHeader, strconv.FormatInt(exemplars, 10))
 }
 
-func convertV2RequestToV1(req *writev2.Request) (cortexpb.PreallocWriteRequest, error) {
+func convertV2RequestToV1(req *cortexpb.PreallocWriteRequestV2) (cortexpb.PreallocWriteRequest, error) {
 	var v1Req cortexpb.PreallocWriteRequest
 	v1Timeseries := make([]cortexpb.PreallocTimeseries, 0, len(req.Timeseries))
 	var v1Metadata []*cortexpb.MetricMetadata
@@ -177,9 +181,9 @@ func convertV2RequestToV1(req *writev2.Request) (cortexpb.PreallocWriteRequest, 
 		v1Timeseries = append(v1Timeseries, cortexpb.PreallocTimeseries{
 			TimeSeries: &cortexpb.TimeSeries{
 				Labels:     cortexpb.FromLabelsToLabelAdapters(lbs),
-				Samples:    convertV2ToV1Samples(v2Ts.Samples),
-				Exemplars:  convertV2ToV1Exemplars(b, symbols, v2Ts.Exemplars),
-				Histograms: convertV2ToV1Histograms(v2Ts.Histograms),
+				Samples:    v2Ts.Samples,
+				Exemplars:  convertV2ToV1Exemplars(&b, symbols, v2Ts.Exemplars),
+				Histograms: v2Ts.Histograms,
 			},
 		})
 
@@ -198,50 +202,27 @@ func convertV2RequestToV1(req *writev2.Request) (cortexpb.PreallocWriteRequest, 
 	return v1Req, nil
 }
 
-func shouldConvertV2Metadata(metadata writev2.Metadata) bool {
-	return !(metadata.HelpRef == 0 && metadata.UnitRef == 0 && metadata.Type == writev2.Metadata_METRIC_TYPE_UNSPECIFIED) //nolint:staticcheck
+func shouldConvertV2Metadata(metadata cortexpb.MetadataV2) bool {
+	return !(metadata.HelpRef == 0 && metadata.UnitRef == 0 && metadata.Type == cortexpb.METRIC_TYPE_UNSPECIFIED) //nolint:staticcheck
 }
 
-func convertV2ToV1Histograms(histograms []writev2.Histogram) []cortexpb.Histogram {
-	v1Histograms := make([]cortexpb.Histogram, 0, len(histograms))
-
-	for _, h := range histograms {
-		v1Histograms = append(v1Histograms, cortexpb.HistogramWriteV2ProtoToHistogramProto(h))
-	}
-
-	return v1Histograms
-}
-
-func convertV2ToV1Samples(samples []writev2.Sample) []cortexpb.Sample {
-	v1Samples := make([]cortexpb.Sample, 0, len(samples))
-
-	for _, s := range samples {
-		v1Samples = append(v1Samples, cortexpb.Sample{
-			Value:       s.Value,
-			TimestampMs: s.Timestamp,
-		})
-	}
-
-	return v1Samples
-}
-
-func convertV2ToV1Metadata(name string, symbols []string, metadata writev2.Metadata) *cortexpb.MetricMetadata {
+func convertV2ToV1Metadata(name string, symbols []string, metadata cortexpb.MetadataV2) *cortexpb.MetricMetadata {
 	t := cortexpb.UNKNOWN
 
 	switch metadata.Type {
-	case writev2.Metadata_METRIC_TYPE_COUNTER:
+	case cortexpb.METRIC_TYPE_COUNTER:
 		t = cortexpb.COUNTER
-	case writev2.Metadata_METRIC_TYPE_GAUGE:
+	case cortexpb.METRIC_TYPE_GAUGE:
 		t = cortexpb.GAUGE
-	case writev2.Metadata_METRIC_TYPE_HISTOGRAM:
+	case cortexpb.METRIC_TYPE_HISTOGRAM:
 		t = cortexpb.HISTOGRAM
-	case writev2.Metadata_METRIC_TYPE_GAUGEHISTOGRAM:
+	case cortexpb.METRIC_TYPE_GAUGEHISTOGRAM:
 		t = cortexpb.GAUGEHISTOGRAM
-	case writev2.Metadata_METRIC_TYPE_SUMMARY:
+	case cortexpb.METRIC_TYPE_SUMMARY:
 		t = cortexpb.SUMMARY
-	case writev2.Metadata_METRIC_TYPE_INFO:
+	case cortexpb.METRIC_TYPE_INFO:
 		t = cortexpb.INFO
-	case writev2.Metadata_METRIC_TYPE_STATESET:
+	case cortexpb.METRIC_TYPE_STATESET:
 		t = cortexpb.STATESET
 	}
 
@@ -253,12 +234,11 @@ func convertV2ToV1Metadata(name string, symbols []string, metadata writev2.Metad
 	}
 }
 
-func convertV2ToV1Exemplars(b labels.ScratchBuilder, symbols []string, v2Exemplars []writev2.Exemplar) []cortexpb.Exemplar {
+func convertV2ToV1Exemplars(b *labels.ScratchBuilder, symbols []string, v2Exemplars []cortexpb.ExemplarV2) []cortexpb.Exemplar {
 	v1Exemplars := make([]cortexpb.Exemplar, 0, len(v2Exemplars))
 	for _, e := range v2Exemplars {
-		promExemplar := e.ToExemplar(&b, symbols)
 		v1Exemplars = append(v1Exemplars, cortexpb.Exemplar{
-			Labels:      cortexpb.FromLabelsToLabelAdapters(promExemplar.Labels),
+			Labels:      cortexpb.FromLabelsToLabelAdapters(e.ToLabels(b, symbols)),
 			Value:       e.Value,
 			TimestampMs: e.Timestamp,
 		})